add error handling for `CDemoEdit` job

+++ src/engine/client/demoedit.cpp
@@ -20,7 +20,8 @@ CDemoEdit::CDemoEdit(const char *pNetVersion, class CSnapshotDelta *pSnapshotDel
 void CDemoEdit::Run()
 {
 	// Slice the current demo
-	m_DemoEditor.Slice(m_aDemo, m_aDst, m_StartTick, m_EndTick, NULL, 0);
-	// We remove the temporary demo file
-	m_pStorage->RemoveFile(m_aDemo, IStorage::TYPE_SAVE);
+	m_Success = m_DemoEditor.Slice(m_aDemo, m_aDst, m_StartTick, m_EndTick, NULL, 0);
+	// We remove the temporary demo file if slicing is successful
+	if(m_Success)
+		m_pStorage->RemoveFile(m_aDemo, IStorage::TYPE_SAVE);
 }
